modified for profiler

+++ src/cd_name_t.h
@@ -95,7 +95,7 @@ public:
     return (level_ == that.level()) && (rank_in_level_ == that.rank_in_level()) && (size_ == that.size());
   }
 
-  uint64_t GetCDName(void) const { 
+  int GetCDName(void) const { 
     return ((level_ << 16) | rank_in_level_);
   }
 };
